constants, size_config, enums moved to helper
