+++ lib/components/form_error.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import '../size_config.dart';
+import '../helper/size_config.dart';
 
 class FormError extends StatelessWidget {
   const FormError({
