+++ lib/components/socal_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import '../size_config.dart';
+import '../helper/size_config.dart';
 
 class SocalCard extends StatelessWidget {
   const SocalCard({
