+++ lib/screens/profile/components/profile_menu.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import '../../../constants.dart';
+import '../../../helper/constants.dart';
 
 class ProfileMenu extends StatelessWidget {
   const ProfileMenu({
