+++ lib/screens/sign_up/components/body.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:plant_shop_app/components/socal_card.dart';
-import 'package:plant_shop_app/constants.dart';
-import 'package:plant_shop_app/size_config.dart';
+import 'package:plant_shop_app/helper/constants.dart';
+import 'package:plant_shop_app/helper/size_config.dart';
 
 import 'sign_up_form.dart';
