+++ lib/screens/sign_up/sign_up_screen.dart
@@ -5,7 +5,7 @@ import 'package:plant_shop_app/components/custom_surfix_icon.dart';
 import 'package:plant_shop_app/components/default_button.dart';
 import 'package:plant_shop_app/components/form_error.dart';
 import 'package:plant_shop_app/components/socal_card.dart';
-import 'package:plant_shop_app/size_config.dart';
+import 'package:plant_shop_app/helper/size_config.dart';
 
 class SignUpScreen extends StatefulWidget {
   @override
